Fix signed/unsigned comparison warning.

+++ src/magnet/metadatadownload.cpp
@@ -71,7 +71,7 @@ namespace bt
 		// piece fits, so copy into data
 		//Out(SYS_GEN|LOG_NOTICE) << "Metadata download, dowloaded " << piece << endl;
 		Uint32 off = piece * METADATA_PIECE_SIZE;
-		if (metadata.size() < off + size)
+		if (static_cast<size_t>(metadata.size()) < off + size)
 		{
 			Out(SYS_GEN|LOG_NOTICE) << "Metadata download, received large piece " << (off + size) << ", max " << metadata.size() << endl;
 			metadata.resize(off + size);
